Move methods CDTOperationRequestBuilder requires to protocol
Essentially, make it so the operation request builder requires
any NSOperation which conforms to CDTOperationRequestBuilderDelegate.

This removes the HTTP request building methods from the
CDTCouchOperation implementation to make it clearer the purpose of the
httpX methods.

The delegate also has the queryItems and rootURL methods, as they
are also required for the builder. These are not removed from
CDTCouchOperation, however, as their setters are used.