Make queryItems a method rather than property
This moves some of the request building logic into the methods
specified by the request builder delegate, rather than in the
-buildAndValidate method.

However, the -buildAndValidate method still does building for some
more complicated requests, where we need to check the validity of
constructed POST bodies. This is to avoid doing the building work
twice during a request.