Add CDTOperationRequestBuilder with one example
The operations in Objective-Cloudant have a lot of copy and paste code
used to build requests. This commit introduces a builder for the
NSURLRequest used to make the request to Cloudant. The idea is to remove
boilerplate code from the operations, in order to bring each operation
closer to the ideal of having only the code it needs and no more.

The builder takes a CDTCouchOperation in its constructor. The
CDTCouchOperation has been updated to expose information required to
build the HTTP request as properties. Subclasses are expected to
override these properties as required for their specific requests.
For example, all operations will override the `httpPath` property.
The builder reads this information into private properties at init
time which are later used to build the request.

It includes changes to CDTPutDocumentOperation which show how this
builder simplifies the code in the operation class.