Move request boilerplate to CDTOperationRequestExecutor
Most of the process of making a request is the same in all operations:
create the request object, create a task, wait for the task, check
cancelled, do some special processing and the complete the operation.

The only part that is specific to each operation is the "do some special
processing" part.

This commit pulls all the other surrounding code into an executor class
whose job is to take an NSOperation implementing
CDTOperationRequestExecutorDelegate and make the request using the
values returned from the delegate instance, handling all the
NSOperation and URL session stuff.