It makes more sense for the executor to be strongly retained
The Operation creates the Executor, so the operation should be the one
strongly retaining the executor. This more closely mirrors the use of
weak references in delegate usage -- the operation should be longer
living than the executor, so should hold the strong reference, just as
the delegate object is typically longer living than the object it is
delegate for.