Add DeleteAttachmentOperation
Added DeleteAttachmentOperation, the properties required for this is a subset of
PutAttachmentOperation, so they now move into a Super class AttachmentOperation, although
said class is not supposed to be used directly it will complete a HEAD request. However
due to the currently implementation of result processing, it will marked as Errored.
This is something that will need to be fixed at a later time, when HEAD operations
become truely part of the API.

Fixes #65