Move to using delegates to enable streaming support
To make it easier to support streaming of data from the server, we need to move
to using delegates so we can dispatch the data back to the requesting operation
as it becomes avaiable. This changes how some things are implemented. URLSessionTask
is now just a wrapper for a NSURLSessionDataTask, said task may be swapped out depending
on the result of the response interceptor run when the response is reccieved.

In the even the response is reccived from the server and the interceptor run
indicates the request should be retired, the task will be cancelled, and removed
from the taskDict and URLSessionTask, and replaced with a new task which represents
the rerun. Assuming the interceptor run allows for the response to be delivered, the
resposne and data will be deliver to the requesting operation via the InterableSessionDelegate,
this makes it possbile to implement streaming outside of the URLSession class making it possible
to provide the same interface to both streaming and non streaming operations.