InterceptableSession Cookie Handling
Add session cookie handling to the InterceptableSession. When the first request for a session is
being made, it will attempt to create a cookie if a username and password have been provided. If
the session reccieves a 401 response for an _session request it will not attempt to get a cookie
from the server again.

If a `Set-Cookie:` header is recieved from the server as part of normal reuqest / response flows,
the cookie will be updated in the cookie store.

Resolves #124