Fix issue where 401 response triggers infinite loop.
Fix issue where 401 response triggers infinite retry loop, this is because
the cookie retry code did not decrement the number of retries remaing for the task.

The tests did not pick this up because 401 responses sent by tests ommited the
"error" key which is required to trigger the handling retry logic.

Fixes #135