Re-organise Source
Re-organise Source so it is split more loggically, this doesn't go as far as creating sub-modules
it appears that sub-modules aren't present with SwiftPM.

The new structure is as follows

Source
|
|-SwiftCloudant
  |
  |------Operations
  |       |
  |       |----Database
  |       |----Documents
  |       |----Query
  |       |----Server
  |       |----Views
  |
  |----- HTTP

The top level SwiftCloudant dir contains The client object and classes relating to that.

Operations contain the Operations which are executed, such as "PutDocumentOperation", this is
split into operationt types, such as `Database` for operation which execute on a database, and `Documents`
for operations which execute on documents.

The `Operations` dir itself, contains the protocols that all operations should implement, and our Foundation.Operation
subclass.

The `HTTP` dir is where the HTTP classes reside, this includes the RequestExecutor, RequestBuilder and URLSession.